backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
supabase==2.18.1
PyJWT==2.10.1
pydantic==2.11.7
ultralytics>=8.3.0,<9.0.0
Pillow>=10.0.0
python-dotenv>=1.0.1
